TST: reorganize misplaced year / business year tests (#27085) (#42930)
