+++ pandas/tests/tseries/offsets/test_yqm_offsets.py
@@ -615,511 +615,3 @@ class TestBQuarterEnd(Base):
     def test_is_on_offset(self, case):
         offset, dt, expected = case
         assert_is_on_offset(offset, dt, expected)
-
-
-# --------------------------------------------------------------------
-# Years
-
-
-class TestYearBegin(Base):
-    _offset = YearBegin
-
-    def test_misspecified(self):
-        with pytest.raises(ValueError, match="Month must go from 1 to 12"):
-            YearBegin(month=13)
-
-    offset_cases = []
-    offset_cases.append(
-        (
-            YearBegin(),
-            {
-                datetime(2008, 1, 1): datetime(2009, 1, 1),
-                datetime(2008, 6, 30): datetime(2009, 1, 1),
-                datetime(2008, 12, 31): datetime(2009, 1, 1),
-                datetime(2005, 12, 30): datetime(2006, 1, 1),
-                datetime(2005, 12, 31): datetime(2006, 1, 1),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            YearBegin(0),
-            {
-                datetime(2008, 1, 1): datetime(2008, 1, 1),
-                datetime(2008, 6, 30): datetime(2009, 1, 1),
-                datetime(2008, 12, 31): datetime(2009, 1, 1),
-                datetime(2005, 12, 30): datetime(2006, 1, 1),
-                datetime(2005, 12, 31): datetime(2006, 1, 1),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            YearBegin(3),
-            {
-                datetime(2008, 1, 1): datetime(2011, 1, 1),
-                datetime(2008, 6, 30): datetime(2011, 1, 1),
-                datetime(2008, 12, 31): datetime(2011, 1, 1),
-                datetime(2005, 12, 30): datetime(2008, 1, 1),
-                datetime(2005, 12, 31): datetime(2008, 1, 1),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            YearBegin(-1),
-            {
-                datetime(2007, 1, 1): datetime(2006, 1, 1),
-                datetime(2007, 1, 15): datetime(2007, 1, 1),
-                datetime(2008, 6, 30): datetime(2008, 1, 1),
-                datetime(2008, 12, 31): datetime(2008, 1, 1),
-                datetime(2006, 12, 29): datetime(2006, 1, 1),
-                datetime(2006, 12, 30): datetime(2006, 1, 1),
-                datetime(2007, 1, 1): datetime(2006, 1, 1),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            YearBegin(-2),
-            {
-                datetime(2007, 1, 1): datetime(2005, 1, 1),
-                datetime(2008, 6, 30): datetime(2007, 1, 1),
-                datetime(2008, 12, 31): datetime(2007, 1, 1),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            YearBegin(month=4),
-            {
-                datetime(2007, 4, 1): datetime(2008, 4, 1),
-                datetime(2007, 4, 15): datetime(2008, 4, 1),
-                datetime(2007, 3, 1): datetime(2007, 4, 1),
-                datetime(2007, 12, 15): datetime(2008, 4, 1),
-                datetime(2012, 1, 31): datetime(2012, 4, 1),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            YearBegin(0, month=4),
-            {
-                datetime(2007, 4, 1): datetime(2007, 4, 1),
-                datetime(2007, 3, 1): datetime(2007, 4, 1),
-                datetime(2007, 12, 15): datetime(2008, 4, 1),
-                datetime(2012, 1, 31): datetime(2012, 4, 1),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            YearBegin(4, month=4),
-            {
-                datetime(2007, 4, 1): datetime(2011, 4, 1),
-                datetime(2007, 4, 15): datetime(2011, 4, 1),
-                datetime(2007, 3, 1): datetime(2010, 4, 1),
-                datetime(2007, 12, 15): datetime(2011, 4, 1),
-                datetime(2012, 1, 31): datetime(2015, 4, 1),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            YearBegin(-1, month=4),
-            {
-                datetime(2007, 4, 1): datetime(2006, 4, 1),
-                datetime(2007, 3, 1): datetime(2006, 4, 1),
-                datetime(2007, 12, 15): datetime(2007, 4, 1),
-                datetime(2012, 1, 31): datetime(2011, 4, 1),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            YearBegin(-3, month=4),
-            {
-                datetime(2007, 4, 1): datetime(2004, 4, 1),
-                datetime(2007, 3, 1): datetime(2004, 4, 1),
-                datetime(2007, 12, 15): datetime(2005, 4, 1),
-                datetime(2012, 1, 31): datetime(2009, 4, 1),
-            },
-        )
-    )
-
-    @pytest.mark.parametrize("case", offset_cases)
-    def test_offset(self, case):
-        offset, cases = case
-        for base, expected in cases.items():
-            assert_offset_equal(offset, base, expected)
-
-    on_offset_cases = [
-        (YearBegin(), datetime(2007, 1, 3), False),
-        (YearBegin(), datetime(2008, 1, 1), True),
-        (YearBegin(), datetime(2006, 12, 31), False),
-        (YearBegin(), datetime(2006, 1, 2), False),
-    ]
-
-    @pytest.mark.parametrize("case", on_offset_cases)
-    def test_is_on_offset(self, case):
-        offset, dt, expected = case
-        assert_is_on_offset(offset, dt, expected)
-
-
-class TestYearEnd(Base):
-    _offset = YearEnd
-
-    def test_misspecified(self):
-        with pytest.raises(ValueError, match="Month must go from 1 to 12"):
-            YearEnd(month=13)
-
-    offset_cases = []
-    offset_cases.append(
-        (
-            YearEnd(),
-            {
-                datetime(2008, 1, 1): datetime(2008, 12, 31),
-                datetime(2008, 6, 30): datetime(2008, 12, 31),
-                datetime(2008, 12, 31): datetime(2009, 12, 31),
-                datetime(2005, 12, 30): datetime(2005, 12, 31),
-                datetime(2005, 12, 31): datetime(2006, 12, 31),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            YearEnd(0),
-            {
-                datetime(2008, 1, 1): datetime(2008, 12, 31),
-                datetime(2008, 6, 30): datetime(2008, 12, 31),
-                datetime(2008, 12, 31): datetime(2008, 12, 31),
-                datetime(2005, 12, 30): datetime(2005, 12, 31),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            YearEnd(-1),
-            {
-                datetime(2007, 1, 1): datetime(2006, 12, 31),
-                datetime(2008, 6, 30): datetime(2007, 12, 31),
-                datetime(2008, 12, 31): datetime(2007, 12, 31),
-                datetime(2006, 12, 29): datetime(2005, 12, 31),
-                datetime(2006, 12, 30): datetime(2005, 12, 31),
-                datetime(2007, 1, 1): datetime(2006, 12, 31),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            YearEnd(-2),
-            {
-                datetime(2007, 1, 1): datetime(2005, 12, 31),
-                datetime(2008, 6, 30): datetime(2006, 12, 31),
-                datetime(2008, 12, 31): datetime(2006, 12, 31),
-            },
-        )
-    )
-
-    @pytest.mark.parametrize("case", offset_cases)
-    def test_offset(self, case):
-        offset, cases = case
-        for base, expected in cases.items():
-            assert_offset_equal(offset, base, expected)
-
-    on_offset_cases = [
-        (YearEnd(), datetime(2007, 12, 31), True),
-        (YearEnd(), datetime(2008, 1, 1), False),
-        (YearEnd(), datetime(2006, 12, 31), True),
-        (YearEnd(), datetime(2006, 12, 29), False),
-    ]
-
-    @pytest.mark.parametrize("case", on_offset_cases)
-    def test_is_on_offset(self, case):
-        offset, dt, expected = case
-        assert_is_on_offset(offset, dt, expected)
-
-
-class TestYearEndDiffMonth(Base):
-    offset_cases = []
-    offset_cases.append(
-        (
-            YearEnd(month=3),
-            {
-                datetime(2008, 1, 1): datetime(2008, 3, 31),
-                datetime(2008, 2, 15): datetime(2008, 3, 31),
-                datetime(2008, 3, 31): datetime(2009, 3, 31),
-                datetime(2008, 3, 30): datetime(2008, 3, 31),
-                datetime(2005, 3, 31): datetime(2006, 3, 31),
-                datetime(2006, 7, 30): datetime(2007, 3, 31),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            YearEnd(0, month=3),
-            {
-                datetime(2008, 1, 1): datetime(2008, 3, 31),
-                datetime(2008, 2, 28): datetime(2008, 3, 31),
-                datetime(2008, 3, 31): datetime(2008, 3, 31),
-                datetime(2005, 3, 30): datetime(2005, 3, 31),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            YearEnd(-1, month=3),
-            {
-                datetime(2007, 1, 1): datetime(2006, 3, 31),
-                datetime(2008, 2, 28): datetime(2007, 3, 31),
-                datetime(2008, 3, 31): datetime(2007, 3, 31),
-                datetime(2006, 3, 29): datetime(2005, 3, 31),
-                datetime(2006, 3, 30): datetime(2005, 3, 31),
-                datetime(2007, 3, 1): datetime(2006, 3, 31),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            YearEnd(-2, month=3),
-            {
-                datetime(2007, 1, 1): datetime(2005, 3, 31),
-                datetime(2008, 6, 30): datetime(2007, 3, 31),
-                datetime(2008, 3, 31): datetime(2006, 3, 31),
-            },
-        )
-    )
-
-    @pytest.mark.parametrize("case", offset_cases)
-    def test_offset(self, case):
-        offset, cases = case
-        for base, expected in cases.items():
-            assert_offset_equal(offset, base, expected)
-
-    on_offset_cases = [
-        (YearEnd(month=3), datetime(2007, 3, 31), True),
-        (YearEnd(month=3), datetime(2008, 1, 1), False),
-        (YearEnd(month=3), datetime(2006, 3, 31), True),
-        (YearEnd(month=3), datetime(2006, 3, 29), False),
-    ]
-
-    @pytest.mark.parametrize("case", on_offset_cases)
-    def test_is_on_offset(self, case):
-        offset, dt, expected = case
-        assert_is_on_offset(offset, dt, expected)
-
-
-class TestBYearBegin(Base):
-    _offset = BYearBegin
-
-    def test_misspecified(self):
-        msg = "Month must go from 1 to 12"
-        with pytest.raises(ValueError, match=msg):
-            BYearBegin(month=13)
-        with pytest.raises(ValueError, match=msg):
-            BYearEnd(month=13)
-
-    offset_cases = []
-    offset_cases.append(
-        (
-            BYearBegin(),
-            {
-                datetime(2008, 1, 1): datetime(2009, 1, 1),
-                datetime(2008, 6, 30): datetime(2009, 1, 1),
-                datetime(2008, 12, 31): datetime(2009, 1, 1),
-                datetime(2011, 1, 1): datetime(2011, 1, 3),
-                datetime(2011, 1, 3): datetime(2012, 1, 2),
-                datetime(2005, 12, 30): datetime(2006, 1, 2),
-                datetime(2005, 12, 31): datetime(2006, 1, 2),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            BYearBegin(0),
-            {
-                datetime(2008, 1, 1): datetime(2008, 1, 1),
-                datetime(2008, 6, 30): datetime(2009, 1, 1),
-                datetime(2008, 12, 31): datetime(2009, 1, 1),
-                datetime(2005, 12, 30): datetime(2006, 1, 2),
-                datetime(2005, 12, 31): datetime(2006, 1, 2),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            BYearBegin(-1),
-            {
-                datetime(2007, 1, 1): datetime(2006, 1, 2),
-                datetime(2009, 1, 4): datetime(2009, 1, 1),
-                datetime(2009, 1, 1): datetime(2008, 1, 1),
-                datetime(2008, 6, 30): datetime(2008, 1, 1),
-                datetime(2008, 12, 31): datetime(2008, 1, 1),
-                datetime(2006, 12, 29): datetime(2006, 1, 2),
-                datetime(2006, 12, 30): datetime(2006, 1, 2),
-                datetime(2006, 1, 1): datetime(2005, 1, 3),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            BYearBegin(-2),
-            {
-                datetime(2007, 1, 1): datetime(2005, 1, 3),
-                datetime(2007, 6, 30): datetime(2006, 1, 2),
-                datetime(2008, 12, 31): datetime(2007, 1, 1),
-            },
-        )
-    )
-
-    @pytest.mark.parametrize("case", offset_cases)
-    def test_offset(self, case):
-        offset, cases = case
-        for base, expected in cases.items():
-            assert_offset_equal(offset, base, expected)
-
-
-class TestBYearEnd(Base):
-    _offset = BYearEnd
-
-    offset_cases = []
-    offset_cases.append(
-        (
-            BYearEnd(),
-            {
-                datetime(2008, 1, 1): datetime(2008, 12, 31),
-                datetime(2008, 6, 30): datetime(2008, 12, 31),
-                datetime(2008, 12, 31): datetime(2009, 12, 31),
-                datetime(2005, 12, 30): datetime(2006, 12, 29),
-                datetime(2005, 12, 31): datetime(2006, 12, 29),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            BYearEnd(0),
-            {
-                datetime(2008, 1, 1): datetime(2008, 12, 31),
-                datetime(2008, 6, 30): datetime(2008, 12, 31),
-                datetime(2008, 12, 31): datetime(2008, 12, 31),
-                datetime(2005, 12, 31): datetime(2006, 12, 29),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            BYearEnd(-1),
-            {
-                datetime(2007, 1, 1): datetime(2006, 12, 29),
-                datetime(2008, 6, 30): datetime(2007, 12, 31),
-                datetime(2008, 12, 31): datetime(2007, 12, 31),
-                datetime(2006, 12, 29): datetime(2005, 12, 30),
-                datetime(2006, 12, 30): datetime(2006, 12, 29),
-                datetime(2007, 1, 1): datetime(2006, 12, 29),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            BYearEnd(-2),
-            {
-                datetime(2007, 1, 1): datetime(2005, 12, 30),
-                datetime(2008, 6, 30): datetime(2006, 12, 29),
-                datetime(2008, 12, 31): datetime(2006, 12, 29),
-            },
-        )
-    )
-
-    @pytest.mark.parametrize("case", offset_cases)
-    def test_offset(self, case):
-        offset, cases = case
-        for base, expected in cases.items():
-            assert_offset_equal(offset, base, expected)
-
-    on_offset_cases = [
-        (BYearEnd(), datetime(2007, 12, 31), True),
-        (BYearEnd(), datetime(2008, 1, 1), False),
-        (BYearEnd(), datetime(2006, 12, 31), False),
-        (BYearEnd(), datetime(2006, 12, 29), True),
-    ]
-
-    @pytest.mark.parametrize("case", on_offset_cases)
-    def test_is_on_offset(self, case):
-        offset, dt, expected = case
-        assert_is_on_offset(offset, dt, expected)
-
-
-class TestBYearEndLagged(Base):
-    _offset = BYearEnd
-
-    def test_bad_month_fail(self):
-        msg = "Month must go from 1 to 12"
-        with pytest.raises(ValueError, match=msg):
-            BYearEnd(month=13)
-        with pytest.raises(ValueError, match=msg):
-            BYearEnd(month=0)
-
-    offset_cases = []
-    offset_cases.append(
-        (
-            BYearEnd(month=6),
-            {
-                datetime(2008, 1, 1): datetime(2008, 6, 30),
-                datetime(2007, 6, 30): datetime(2008, 6, 30),
-            },
-        )
-    )
-
-    offset_cases.append(
-        (
-            BYearEnd(n=-1, month=6),
-            {
-                datetime(2008, 1, 1): datetime(2007, 6, 29),
-                datetime(2007, 6, 30): datetime(2007, 6, 29),
-            },
-        )
-    )
-
-    @pytest.mark.parametrize("case", offset_cases)
-    def test_offset(self, case):
-        offset, cases = case
-        for base, expected in cases.items():
-            assert_offset_equal(offset, base, expected)
-
-    def test_roll(self):
-        offset = BYearEnd(month=6)
-        date = datetime(2009, 11, 30)
-
-        assert offset.rollforward(date) == datetime(2010, 6, 30)
-        assert offset.rollback(date) == datetime(2009, 6, 30)
-
-    on_offset_cases = [
-        (BYearEnd(month=2), datetime(2007, 2, 28), True),
-        (BYearEnd(month=6), datetime(2007, 6, 30), False),
-    ]
-
-    @pytest.mark.parametrize("case", on_offset_cases)
-    def test_is_on_offset(self, case):
-        offset, dt, expected = case
-        assert_is_on_offset(offset, dt, expected)
